+++ packages/zarm/src/index.ts
@@ -10,7 +10,8 @@ export { default as Affix } from './affix';
 export type { AlertProps } from './alert';
 // export { default as Alert } from './alert';
 
-export { default as BackToTop } from './back-to-top';
+export type { BackTopProps } from './back-top';
+export { default as BackTop } from './back-top';
 
 export type { BadgeProps } from './badge';
 export { default as Badge } from './badge';